Fixed some issues

+++ Dao/src/main/java/me/exrates/dao/WalletDao.java
@@ -1,5 +1,6 @@
 package me.exrates.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import me.exrates.model.Wallet;
@@ -10,9 +11,9 @@ public interface WalletDao {
 
 	double getWalletRBalance(int walletId);
 
-	boolean setWalletABalance(int walletId,double newBalance);
+	boolean setWalletABalance(int walletId, BigDecimal newBalance);
 
-	boolean setWalletRBalance(int walletId,double newBalance);
+	boolean setWalletRBalance(int walletId, BigDecimal newBalance);
 
 	int getWalletId(int userId, int currencyId);
